% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovidMali-package.R
\docType{package}
\name{CovidMali2-package}
\alias{CovidMali2}
\alias{CovidMali2-package}
\title{CovidMali2: Covid in Mali}
\description{
Mali Covid cases
}
\author{
\strong{Maintainer}: Patrick Kelly \email{phototala@mail.com}

}
\keyword{internal}
